refactor: move type definitions so that 'zod' isn't polluting the client bundle
- The '@/shared' folder should only ever import from the '@/shared' folder, so move around types and definitions to deal with this
- Move any 'zod'ified consents into its own file, so that when importing the type on the client zod isn't included

This saves around 13kb gzipped!
