+++ src/server/lib/okta/register.ts
@@ -1,8 +1,4 @@
-import {
-	RegistrationLocation,
-	Status,
-	UserResponse,
-} from '@/server/models/okta/User';
+import { Status, UserResponse } from '@/server/models/okta/User';
 import {
 	createUser,
 	getUser,
@@ -24,8 +20,9 @@ import { logger } from '../serverSideLogger';
 import dangerouslySetPlaceholderPassword from './dangerouslySetPlaceholderPassword';
 import { sendCompleteRegistration } from '@/email/templates/CompleteRegistration/sendCompleteRegistration';
 import { addAppPrefixToOktaRecoveryToken } from '@/server/lib/deeplink/oktaRecoveryToken';
-import { RegistrationConsents } from '@/shared/model/Consent';
 import { encryptRegistrationConsents } from '@/server/lib/registrationConsents';
+import { RegistrationConsents } from '@/shared/model/RegistrationConsents';
+import { RegistrationLocation } from '@/shared/model/RegistrationLocation';
 
 const { okta } = getConfiguration();
