+++ src/shared/lib/newsletter.ts
@@ -1,4 +1,5 @@
-import { ALL_NEWSLETTER_IDS, NewsletterPatch } from '@/shared/model/Newsletter';
+import { ALL_NEWSLETTER_IDS } from '@/shared/model/Newsletter';
+import { NewsletterPatch } from '@/shared/model/NewsletterPatch';
 
 // get a list of newsletters that have been updated in the body and compare
 // to list of all newsletter ids
