+++ src/shared/model/Newsletter.ts
@@ -1,11 +1,3 @@
-import { z } from 'zod';
-
-export const newsletterPatchSchema = z.object({
-	id: z.string(),
-	subscribed: z.boolean(),
-});
-export type NewsletterPatch = z.infer<typeof newsletterPatchSchema>;
-
 export interface NewsLetter {
 	id: string;
 	nameId: string;
