+++ src/shared/model/NewsletterPatch.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+
+export const newsletterPatchSchema = z.object({
+	id: z.string(),
+	subscribed: z.boolean(),
+});
+export type NewsletterPatch = z.infer<typeof newsletterPatchSchema>;
