+++ src/shared/model/User.ts
@@ -1,11 +1,5 @@
-import { z } from 'zod';
-import { RegistrationLocation } from '@/server/models/okta/User';
-
-export const userConsentSchema = z.object({
-	id: z.string(),
-	consented: z.boolean().optional(),
-});
-export type UserConsent = z.infer<typeof userConsentSchema>;
+import { UserConsent } from '@/shared/model/UserConsents';
+import { RegistrationLocation } from '@/shared/model/RegistrationLocation';
 
 export default interface User {
 	consents: UserConsent[];
